create pages directory and moved Home and Signup to it.
