+++ src/main.tsx
@@ -5,8 +5,8 @@ import { createBrowserRouter, RouterProvider } from 'react-router';
 
 import './index.css';
 
-import Home from '@components/Home';
-import Signup from '@components/Signup';
+import Home from 'src/pages/Home';
+import Signup from 'src/pages/Signup';
 import NotFound from '@components/common/NotFound';
 
 const router = createBrowserRouter([
